python-yuz-servisi/requirements.txt
# Python 3.12 uyumlu bağımlılıklar
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
python-dotenv>=1.0.0

# face_recognition için dlib gerekli
# Windows'ta dlib kurulumu için önce Visual Studio Build Tools gerekli
# veya pip install dlib --prefer-binary kullanılabilir
dlib>=19.24.2
face_recognition>=1.3.0
